fix(waifu): family theater lock, not flag moles
Tip 4c96d571 still substring-matched ignore+pattern on any
argv token (path runs soft-denied) and babysat one flag at
a time. Caps cannot hold that train.

Same path. New families part: flag-shaped only; collection
prefixes (--testpathignore / --modulepathignore + kebab);
pytest --ignore / --ignore-glob (not --ignore-pattern);
--exclude / --dir / --filter families. JS hosts + peeled
test:* share one failed-only and one suite set. Workspace
stays JS-only (cargo clippy --workspace is full). Coverage /
transform / watch-FS ignores and path tokens stay full.
Selectors stay full. VIP keepers unchanged.
